Reorganize components of `sym::vec::` module into submodules `convert` and `ops`
